+++ TRS_Application_EJB/ejbModule/be/iris/session/ActivityBean.java
@@ -22,7 +22,7 @@ public class ActivityBean implements ActivityBeanRemote {
 	}
 
 	@Override
-	public void saveNewActivitie(Tutactivity activity, String pid, int pno) {
+	public void saveNewActivitie(Tutactivity activity, String pid, long pno) {
 		activityDao.insertActivity(activity, pid, pno);
 
 	}
